# Optimize circumsphere and triangulation.py