--- a/adaptive/tests/unit/test_triangulation.py
+++ b/adaptive/tests/unit/test_triangulation.py
@@ -60,3 +60,36 @@ def test_triangulation_find_opposing_vertices_raises_if_simplex_is_invalid():
 
     with pytest.raises(ValueError):
         tri.get_opposing_vertices((2, 3, 5))
+
+
+def test_circumsphere():
+    """ Test that circumsphere works correctly for a random center and random points on a sphere """
+    from adaptive.learner.triangulation import circumsphere, fast_norm
+    from numpy import allclose
+    from numpy.random import normal, uniform
+    center_diff_err = "Calculated center [%s] differs from true center [%s]\n"
+
+    def generate_random_sphere_points(dim, radius=0):
+        """ Refer to https://math.stackexchange.com/a/1585996 """
+
+        vec = [None] * (dim + 1)
+        center = uniform(-100, 100, dim)
+        radius = uniform(1.0, 100.0) if radius == 0 else radius
+        for i in range(dim + 1):
+            points = normal(0, size=dim)
+            x = fast_norm(points)
+            points = points / x * radius
+            vec[i] = tuple(points + center)
+
+        return radius, center, vec
+
+    for dim in range(2, 10):
+        radius, center, points = generate_random_sphere_points(dim)
+        circ_center, circ_radius = circumsphere(points)
+        err_msg = ""
+        if not allclose(circ_center, center):
+            err_msg += center_diff_err % (",".join([str(x) for x in circ_center]), ",".join([str(x) for x in center]))
+        if not allclose(radius, circ_radius):
+            err_msg += "Calculated radius %s differs from true radius %s" % (circ_radius, radius)
+        if err_msg:
+            raise AssertionError(err_msg)