--- a/adaptive/tests/unit/test_triangulation.py
+++ b/adaptive/tests/unit/test_triangulation.py
@@ -63,13 +63,12 @@ def test_triangulation_find_opposing_vertices_raises_if_simplex_is_invalid():
 
 
 def test_circumsphere():
-    from adaptive.learner.triangulation import circumsphere
+    from adaptive.learner.triangulation import circumsphere, fast_norm
     from numpy import allclose
+    from numpy.random import normal, uniform
 
     def generate_random_sphere_points(dim, radius=0):
         """ Refer to https://math.stackexchange.com/a/1585996 """
-        from numpy.random import normal, uniform
-        from adaptive.learner.triangulation import fast_norm
 
         vec = [None] * (dim + 1)
         center = uniform(-100, 100, dim)
@@ -82,15 +81,15 @@ def test_circumsphere():
 
         return radius, center, vec
 
-    center_diff_err = "Calculated center [%s] differs from true center [%s]\n"
+    center_diff_err = "Calculated center (%s) differs from true center (%s)\n"
     for dim in range(2, 10):
         radius, center, points = generate_random_sphere_points(dim)
         circ_center, circ_radius = circumsphere(points)
         err_msg = ""
         if not allclose(circ_center, center):
             err_msg += center_diff_err % (
-                ",".join([str(x) for x in circ_center]),
-                ",".join([str(x) for x in center]),
+                ", ".join([str(x) for x in circ_center]),
+                ", ".join([str(x) for x in center]),
             )
         if not allclose(radius, circ_radius):
             err_msg += "Calculated radius {} differs from true radius {}".format(