@basnijholt 2019-12-16
Thanks a lot for doing these optimizations!

You can easily fix all the style issues by running `pre-commit run --all` (first install using `pip install pre-commit`).

I've made https://github.com/python-adaptive/adaptive/pull/246 to make this more obvious in the future.