@akhmerov 2019-12-16
Should we use `@` instead of `dot`? I might be wrong, but it seems either faster or same speed, and more readable.