@akhmerov 2019-12-16
While most of the names that we import have no collision with built-ins, `abs` does override the built-in one. Would it be a good idea to import `abs` as `np_abs`?

Is making all names available in the module namespace an important optimization?