@philippeitis 2019-12-16
> While most of the names that we import have no collision with built-ins, `abs` does override the built-in one. Would it be a good idea to import `abs` as `np_abs`?
> 
> Is making all names available in the module namespace an important optimization?

As a quick demo:

>python -m timeit -s"from math import sqrt" "sqrt(5.6)"
5000000 loops, best of 5: 69.4 nsec per loop

>python -m timeit -s"import math" "math.sqrt(5.6)"
5000000 loops, best of 5: 94.8 nsec per loop

A 20-30% performance boost from direct imports is pretty nice to have, though this may vary for specific functions and their overhead.