@philippeitis 2019-12-16
> Should we use `@` instead of `dot`? I might be wrong, but it seems either faster or same speed, and more readable.

The issue with dot is that it doesn't work if the argument is a regular list. Also, it seems to be slower, at least from my testing:
```
>python -m timeit -s"from numpy import array;vect = array([1,2,3, 4])" "vect @ vect"
500000 loops, best of 5: 879 nsec per loop

>python -m timeit -s"from numpy import dot, array; vect = array([1,2,3,4])" "dot(vect, vect)"
500000 loops, best of 5: 603 nsec per loop
```