@philippeitis 2019-12-16
Azure seems to be throwing this error which is causing the linting tests to fail:
```
Traceback (most recent call last):
  File "/home/vsts/.cache/pre-commit/repom1wn1zn2/py_env-python3.7/bin/black", line 5, in <module>
    from black import patched_main
  File "/home/vsts/.cache/pre-commit/repom1wn1zn2/py_env-python3.7/lib/python3.7/site-packages/black.py", line 15, in <module>
    import regex as re
  File "/home/vsts/.cache/pre-commit/repom1wn1zn2/py_env-python3.7/lib/python3.7/site-packages/regex.py", line 402, in <module>
    import _regex_core
  File "/home/vsts/.cache/pre-commit/repom1wn1zn2/py_env-python3.7/lib/python3.7/site-packages/_regex_core.py", line 21, in <module>
    import regex._regex as _regex
ModuleNotFoundError: No module named 'regex._regex'; 'regex' is not a package
```