@basnijholt 2019-12-16
That failure is because of https://github.com/psf/black/issues/1207.

I imagine that will be fixed soon.