@jbweston 2020-05-05
I merged this manually (so that we could get a rebase but also an explicit merge commit, which GitHub does not allow, for some reason), but it seems that GitHub did not register this as closing this PR.

@philippeitis your changes are merged in, but I will have to "close" this PR.