@jbweston 2020-05-05
@philippeitis Let me know how you'd like me to credit you in the AUTHORS file (your GH profile does not have your real name).