@philippeitis 2020-05-08
Happy to see these changes merged in. If you'd like to credit me in the AUTHORS file, my name is Philippe Solodov.